server_side/requirements.txt
flask==5.1.0
flask-cors==5.0.0
flask-socketio==5.3.6
python-dotenv==1.0.1
twilio==5.7.0